Move plugins to its own file
